html/semantics/text-level-semantics/the-data-element/001.html

<!doctype html>
<html>
  <head>
    <meta charset=utf-8>
    <title>HTML data element API</title>
    <style>
data { visibility: hidden; }
    </style>
    <link rel="help" href="https://html.spec.whatwg.org/multipage/#the-data-element">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <!-- intentionally nested to test parsing rules -->
    <p id="data"><data value="Dummy text">Dummy text <data>Other text<data>more text<data></data></data>even more</data></data></p>
    <script type="text/javascript">
function makeData(value,contents,valueProp) {
  var dataEl = document.createElement('data');
  if( value ) {
    dataEl.setAttribute('value',value);
  }
  if( contents ) {
    dataEl.innerHTML = contents;
  }
  if( valueProp ) {
    dataEl.value = valueProp;
  }
  return dataEl;
}

var datap = document.getElementById('data');
var datas = datap.getElementsByTagName('data');

//DATA elements
test(function () {
  assert_equals( datas.length, 4 );
}, 'HTML parsing should locate 4 data elements in this document');
test(function () {
  assert_true( !!window.HTMLDataElement );
}, 'HTMLDataElement should be exposed for prototyping');
test(function () {
  assert_true( makeData() instanceof window.HTMLDataElement, 'createElement variant' );
  assert_true( datas[0] instanceof window.HTMLDataElement, 'HTML parsing variant' );
}, 'the data elements should be instanceof HTMLDataElement');

//value
test(function () {
  assert_equals( makeData('Some text','Some text').value, 'Some text' );
}, 'the value attribute should be reflected by the .value property');

test(function () {
  assert_equals( typeof makeData().value, 'string', 'typeof test' );
  assert_equals( makeData().value, '', 'value test' );
}, 'the value IDL property should default to an empty string');

test(function () {
  assert_equals( makeData(false,false,'Some text').value, 'Some text' );
}, 'the value property should be read/write');

test(function () {
  assert_equals( makeData('go fish').value, 'go fish' );
}, 'the value attribute should be reflected by the .value property even if it is invalid');

test(function () {
  assert_equals( makeData(false,'Some text').value, '' );
}, 'the value attribute should not reflect the textContent');

    </script>
  </body>
</html>
